Refactored getting the hospital distances with refresh
