+++ Backend/EMS_Django_Backend/requirements.txt
@@ -1,5 +1,18 @@
 asgiref==3.3.1
+beautifulsoup4==4.9.3
+bs4==0.0.1
+certifi==2020.12.5
+chardet==4.0.0
 Django==3.1.6
+idna==2.10
+lxml==4.6.3
+numpy==1.20.1
+pandas==1.2.3
+python-dateutil==2.8.1
 pytz==2021.1
+requests==2.25.1
+six==1.15.0
+soupsieve==2.2.1
 sqlparse==0.4.1
+urllib3==1.26.4
 wheel==0.24.0
