+++ EMS iOS App/EMS iOS App/UI_Layer/VCs/LoadingVC.swift
@@ -13,57 +13,13 @@ class LoadingVC: UIViewController {
     
     @IBOutlet var spinner: UIActivityIndicatorView!
     var hospitals: Array<HospitalIH>!
-    var locationManager: CLLocationManager!
-
-    func getHospitalDistances() {
-        let currentLocation: CLLocation = self.determineCurrentLocation()
-        var count = 0
-        for hospital in self.hospitals {
-            let hospitalLocation = CLLocation(latitude: CLLocationDegrees(exactly: hospital.lat)!, longitude: CLLocationDegrees(exactly: hospital.long)!)
-            let request = MKDirections.Request()
-            request.source = MKMapItem(placemark: MKPlacemark(coordinate: currentLocation.coordinate))
-            request.destination = MKMapItem(placemark: MKPlacemark(coordinate: hospitalLocation.coordinate))
-            request.requestsAlternateRoutes = true
-            request.transportType = .automobile
-            let directions = MKDirections(request: request)
-            directions.calculate { response, error in
-                var distance = currentLocation.distance(from: hospitalLocation)
-                if let unwrappedResponse = response {
-                    let route = unwrappedResponse.routes[0]
-                    distance = route.distance
-                }
-                distance = (distance / 1000) * 0.62137
-                distance = Double(round(distance * 100) / 100)
-                hospital.distance = distance
-                count += 1
-                if count == self.hospitals.count {
-                    DispatchQueue.main.async {
-                        self.spinner.stopAnimating()
-                        self.performSegue(withIdentifier: "showHospitals", sender: self)
-                    }
-                }
-            }
-        }
-    }
-    
-    func determineCurrentLocation() -> CLLocation {
-        if (CLLocationManager.authorizationStatus() == .authorizedWhenInUse || CLLocationManager.authorizationStatus() == .authorizedAlways){
-            if let location = self.locationManager.location {
-                return location
-            } else {
-                return self.determineCurrentLocation()
-            }
-        }
-        return CLLocation()
-    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
         self.spinner.startAnimating()
-        self.locationManager = CLLocationManager()
-        self.locationManager.requestWhenInUseAuthorization()
-        self.hospitals = Array<HospitalIH>()
         let tasker = HospitalsTasker()
+        tasker.locationManager.requestWhenInUseAuthorization()
+        self.hospitals = Array<HospitalIH>()
         tasker.getAllHospitals(failure: {
             print("Failure")
         }, success: { (hospitals) in
@@ -76,7 +32,11 @@ class LoadingVC: UIViewController {
                 return
             }
             self.hospitals = hospitals
-            self.getHospitalDistances()
+            tasker.getHospitalDistances(hospitals: self.hospitals, finished: { (hospitals) in
+                self.hospitals = hospitals
+                self.spinner.stopAnimating()
+                self.performSegue(withIdentifier: "showHospitals", sender: self)
+            })
         })
     }
